Feature/authorizations

+++ resources/views/projects/edit.blade.php
@@ -13,9 +13,11 @@
                     <strong>{{ $project->name }}</strong>
 
                     <div class="pull-right">
+			@can('delete', $project)
                         {!! Form::open(['method' => 'delete', 'route' => ['projects.destroy', 'project' => $project->slug]]) !!}
                         <button class="btn btn-danger btn-xs" name="delete-resource" type="submit" value="delete">delete</button>
                         {!! Form::close() !!}
+			@endcan
                     </div>
                 </div>
